requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.0
yt-dlp>=2023.10.13
python-multipart>=0.0.6
python-dotenv>=1.0.0
Pillow>=10.0.0
wand>=0.6.11